Parser::parseFunctionInfo hits RELEASE_ASSERT for Arrow Functions https://bugs.webkit.org/show_bug.cgi?id=147090
Reviewed by Yusuke Suzuki.

Source/JavaScriptCore:

ArrowFunction's have there ParserFunctionInfo "name" field to
be a non-null pointer. This is obviously allowed and valid except we
had a RELEASE_ASSERT that claimed otherwise. This is a mistake.

Note: ArrowFunction's will never actually have a function name;
there ParserFunctionInfo "name" field will be the empty string.
This is not be mistaken with the name field being a null pointer.

* parser/Parser.cpp:
(JSC::Parser<LexerType>::parseFunctionInfo):

LayoutTests:

* js/arrowfunction-strict-mode-expected.txt: Added.
* js/arrowfunction-strict-mode.html: Added.
* js/script-tests/arrowfunction-strict-mode.js: Added.
(foo):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187014 268f45cc-cd09-0410-ab3c-d52691b4dbfc