REGRESSION(186691): OSR entry is broken on loop headers that have no live variables https://bugs.webkit.org/show_bug.cgi?id=147074 rdar://problem/21869970
Reviewed by Michael Saboff.
        
The OSR entry must-handle block/value widening introduced in r186691 would cause the
CFA to reexecute if it caused any live local variables to change value. But this fails
if the must-handle block has no live local variables, and the entry block otherwise
appears to be unreachable.
        
This fixes the bug by having the change detection include whether the block hadn't been
visited in addition to whether any local variable values got widened.
        
This is a ~4% speed-up on SunSpider in browser.

* dfg/DFGCFAPhase.cpp:
(JSC::DFG::CFAPhase::run):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187028 268f45cc-cd09-0410-ab3c-d52691b4dbfc