Web Inspector: console.assert(false, "Message") message is not visible in console https://bugs.webkit.org/show_bug.cgi?id=147130
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-07-20
Reviewed by Timothy Hatcher.

* Localizations/en.lproj/localizedStrings.js:
* UserInterface/Views/ConsoleMessageView.js:
(WebInspector.ConsoleMessageView.prototype._appendMessageTextAndArguments):
We were forgetting to assign the result of args.concat back into args.
In re-addressing this, improve the formatting of output if there is
a string message or not as the second argument to console.assert(), as
that is the common usage.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187052 268f45cc-cd09-0410-ab3c-d52691b4dbfc