vmmap crash at JavaScriptCore: 0x31cd12f6 (the JavaScript malloc zone enumerator) https://bugs.webkit.org/show_bug.cgi?id=147274
Reviewed by Anders Carlsson.

It's not really clear why vmmap sometimes fails to read the target
process, but we can avoid a crash when it does. This is useful because
you'll still get all the non-bmalloc data out of the target process,
and bmalloc might not even be relevant to your investigation.

* bmalloc/Zone.cpp:
(bmalloc::remoteRead): Check for failure.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187362 268f45cc-cd09-0410-ab3c-d52691b4dbfc