REGRESSION (r188325): Web Inspector: Fix vertical spacing in CodeMirror https://bugs.webkit.org/show_bug.cgi?id=147971
r188325 inceased line-height by 2px. Remove top and bottom 1px padding
to compensate for line-height changes.

In the feature we may highlight the backgroud of text tokens (e.g. for the
heatmap profiler) so we would want to get rid of the gaps between the lines
(caused by the paddind) regardless of this regression.

Reviewed by Timothy Hatcher.

* UserInterface/Views/CodeMirrorOverrides.css:
(.CodeMirror pre):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188382 268f45cc-cd09-0410-ab3c-d52691b4dbfc