[Win CMake] Fix incremental build after r188673 https://bugs.webkit.org/show_bug.cgi?id=148539
Reviewed by Brent Fulgham.

Source/JavaScriptCore:

* PlatformWin.cmake:
Use xcopy as a build step instead of file(COPY ...) to copy updated headers.

Source/WebCore:

* CMakeLists.txt:
* PlatformWin.cmake:
Use xcopy as a build step instead of file(COPY ...) to copy updated headers.

Tools:

* TestWebKitAPI/PlatformWin.cmake:
Copy forwarding headers from WebCore before building in case that was not yet done.
It used to be done when running CMake, but now it's a build step.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@189055 268f45cc-cd09-0410-ab3c-d52691b4dbfc