WKWebView scroll deceleration rate is not being set properly https://bugs.webkit.org/show_bug.cgi?id=149145 <rdar://problem/22064071>
Reviewed by Simon Fraser.

We are not currently setting the web view's deceleration rate to the correct preferred value because we
are setting the decelerationRate property of the scrollView, which snaps input values to the closer of
two predefined UIKit constants, UIScrollViewDecelerationRateFast and UIScrollViewDecelerationRateNormal.
To fix this, we directly set the horizontal and vertical scroll deceleration factors to the preferred
value.

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView scrollViewWillBeginDragging:]): Directly set the scroll view's deceleration factor without
        going through the decelerationRate property.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@189773 268f45cc-cd09-0410-ab3c-d52691b4dbfc