REGRESSION (r182449, Mavericks ONLY): Pages re-open empty after swiping back and scrolling on them https://bugs.webkit.org/show_bug.cgi?id=149317 <rdar://problem/22521514>
Reviewed by Tim Horton.

Source/WebCore:

Disable on Mavericks a PageCache optimization from r182449 which lets
into PageCache pages that only have certain types of pending loads
(images and XHR). This is because it has been determined via bisection
that this change is the one that introduced the bug on Mavericks.

* loader/DocumentLoader.cpp:
(WebCore::areAllLoadersPageCacheAcceptable):

LayoutTests:

Skip a couple of PageCache layout tests on Mavericks now that a PageCache
optimization has been disabled.

* platform/mac/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@189976 268f45cc-cd09-0410-ab3c-d52691b4dbfc