REGRESSION (r182648): UIWebView does not set location.hash synchronously https://bugs.webkit.org/show_bug.cgi?id=149415 <rdar://problem/22169467>
Reviewed by Darin Adler.

The problem was that the delegate in UI thread became asynchronous due to calling
into LaunchServices.

* WebCoreSupport/WebFrameLoaderClient.mm: (shouldTryAppLink): Don't attempt app
links for same domain navigations. This check is part of substantially more complicated
logic in WebKit2 that I don't understand, but that should be enough to demonstrate
that this fix is an improvement at least.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190092 268f45cc-cd09-0410-ab3c-d52691b4dbfc