Update HTMLTableRowElement.rowIndex to behave according to the specification https://bugs.webkit.org/show_bug.cgi?id=148842 <rdar://problem/22586914>
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

Rebaseline W3C HTML test now that one more check is passing.

* web-platform-tests/html/semantics/tabular-data/the-tr-element/rowIndex-expected.txt:

Source/WebCore:

Update HTMLTableRowElement.rowIndex to behave according to the
specification:
https://html.spec.whatwg.org/multipage/tables.html#dom-tr-rowindex

In particular, rowIndex no longer returns -1 for <tr> elements that are
direct children of the table Element. Also,<tr> Elements in all <thead> /
<tfoot> children of the table Element now get a rowIndex, as per the
specification:
- https://html.spec.whatwg.org/multipage/tables.html#dom-table-rows

Previously, only <tr> elements in the first <thead> and the first <tbody>
children of the table would get a rowIndex.

Also, we now abort early and return -1 if the row is a grand-child of a
table element but its direct parent is not a <thead> / <tbody> / <tfoot>.
This avoid unnecessary traversal and matches the specification.

The code is refactored a bit to leverage HTMLTableElement.rows(), as
per the specification:
"return the index of the tr element in that table element's rows collection"

Previously, we would duplicate the table traversal logic. This
simplifies the code and fixes a bug as the traversal logic was slightly
different here and in HTMLTableRowsCollection. Because of this, only
<tr> elements in the first <thead> / <tfoot> would get a rowIndex.
Performance should be similar.

No new tests, already covered by existing test.

* html/HTMLTableRowElement.cpp:
(WebCore::HTMLTableRowElement::rowIndex):
(WebCore::HTMLTableRowElement::sectionRowIndex): Deleted.
(WebCore::HTMLTableRowElement::insertCell): Deleted.

LayoutTests:

Rebaseline existing test because rows in all <thead> / <tfoot> children
of the <table> Element now get a rowIndex.

* platform/mac/fast/table/rowindex-expected.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190099 268f45cc-cd09-0410-ab3c-d52691b4dbfc