Range.intersectsNode() does not behave according to the specification when start / end are equal https://bugs.webkit.org/show_bug.cgi?id=148774 <rdar://problem/22571418>
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

Rebaseline W3C Range test now that another check is passing.

* web-platform-tests/dom/ranges/Range-intersectsNode-expected.txt:

Source/WebCore:

Range.intersectsNode() does not behave according to the specification
when start / end are equal:
https://dom.spec.whatwg.org/#dom-range-intersectsnode

The specification says:
"If (parent, offset) is before end and (parent, offset + 1) is after
start, return true."

However, the way we were detecting "is before end" and "is after start"
was flawed in the case where start / end were equal. This is because when
comparePoint() return 0, then it means that the node/offset is not only
equal to start but also equal to end. As such, it is not "after start"
/ "before end".

No new tests, already covered by existing.

* dom/Range.cpp:
(WebCore::Range::intersectsNode):
(WebCore::highestAncestorUnderCommonRoot): Deleted.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190136 268f45cc-cd09-0410-ab3c-d52691b4dbfc