Source/WebCore: An SVG inherits the container size of the previously drawn HTMLImageElement when drawing it on a canvas https://bugs.webkit.org/show_bug.cgi?id=148845
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2015-09-28
Reviewed by Darin Adler.

When drawing an SVG image on a canvas we need to make sure that the container
size of the SVG is set to the image source size. If we do not do that, the
previous container size will be used as the size of the image. When drawing
it as HTMLImageElement, the renderer calls SVGImage::drawForContainer() which
calls setContainerSize() and then calls draw(). This call ensures that the
size is set correctly for cached images. But when drawing the SVGImage on
CanvasRenderingContext2D, the SVGImage::draw() is called directly because
it calls it through the base class method Image::draw().

If the image.isSVGImage(), CanvasRenderingContext2D::drawImage() has to call
SVGImage::setContainerSize() before calling Image::draw().

Test: svg/as-image/svg-cached-draw-canvas.html

* html/canvas/CanvasRenderingContext2D.cpp:
(WebCore::CanvasRenderingContext2D::drawImage):
Call SVGImage::setContainerSize() before calling Image::draw().

LayoutTests:
An SVG inherits the container size of the previously drawn HTMLImageElement when drawing it on canvas
https://bugs.webkit.org/show_bug.cgi?id=148845

Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2015-09-28
Reviewed by Darin Adler.

Ensure that an SVG image with intrinsic size or with no intrinsic size can be
displayed on a canvas even if it is previously displayed by an HTMLImageElement.

* svg/as-image/svg-cached-draw-canvas-expected.html: Added.
* svg/as-image/svg-cached-draw-canvas.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190274 268f45cc-cd09-0410-ab3c-d52691b4dbfc