Tile map shows a green rect when threaded scrolling is disabled https://bugs.webkit.org/show_bug.cgi?id=149716
Reviewed by Darin Adler.

Green is supposed to indicate that we're using the fast path; if threaded
scrolling is disabled, we're definitely not doing that.

* platform/graphics/TiledBacking.h:
* platform/graphics/ca/TileController.cpp:
(WebCore::TileController::TileController):
* platform/graphics/ca/TileCoverageMap.cpp:
(WebCore::TileCoverageMap::update):
Default to the "we have no ScrollingCoordinator" purple indication;
if a ScrollingCoordinator comes along it will setScrollingModeIndication
and change it from this default.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190628 268f45cc-cd09-0410-ab3c-d52691b4dbfc