Add a page that cycles through v2 dashboards https://bugs.webkit.org/show_bug.cgi?id=149907
Reviewed by Chris Dumez.

Add cycler.html that goes through each dashboard on v2 UI.

This allows the dashboards to be cycled through on a TV screen.

* public/cycler.html: Added.
(loadURLAt): Appends a new iframe to load the next URL (i is the index of the dashboard to be shown)
at the end of body. We don't immediately show the new iframe since it might take a while to load.
(showNewFrameIfLoaded): Remove the current iframe and show the next iframe if the next dashboard has
finished loading. We can't rely on DOMContentLoaded or load events because we use asynchronous XHR to
load each chart's data. Instead, wait until some chart becomes available or fails to load and none of
charts are still in progress to be shown.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190702 268f45cc-cd09-0410-ab3c-d52691b4dbfc