ASSERT(m_motionManager) on emgn.com page https://bugs.webkit.org/show_bug.cgi?id=150070 <rdar://problem/18383732>
Reviewed by Simon Fraser.

In the WebCoreMotionManager init method, we call
into another setup method on the main thread.
However, if a listener was attached before that
ran, we'd ASSERT. This wasn't actually causing a bug
because the main thread initialization would
check the listeners once it got a chance to run.

The fix is to only check status when we've
actually initialized.

* platform/ios/WebCoreMotionManager.h: New m_initialized boolean.
* platform/ios/WebCoreMotionManager.mm: Check m_initialized before doing real work.
(-[WebCoreMotionManager init]):
(-[WebCoreMotionManager addMotionClient:]):
(-[WebCoreMotionManager removeMotionClient:]):
(-[WebCoreMotionManager addOrientationClient:]):
(-[WebCoreMotionManager removeOrientationClient:]):
(-[WebCoreMotionManager initializeOnMainThread]):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191007 268f45cc-cd09-0410-ab3c-d52691b4dbfc